refactor: show all pagination data
